Database Implementation, Not Successfully Logging In
Created a simple database that can be used for the storing of user information. Wrote all the plumbing for the database connection and have confirmed it is connecting to the database properly, however there is an issue with the logic that controls closing the login window and opening the main application. Originally I was going to do all the subscription work manually, but I've decided to pull in the reactive libraries and do it that way. This commit is before any of that rework as I fear I'm about to open a can of worms that will take a lot of time, effort and tears to close.