Refactoring and Extracting
Realized that the current state of the code was a bit confusing. Repositories weren't really... repositories and there was things like authentication being handled within them which I didn't like. To fix this I've taken the time to abstract some things out and create an actual repository system within the project as I think it'll be useful for later on. I also added in a place holder LoginModule, but I'm not ready to tackle that quite yet. Eventually the goal would be to create a full module system so the project can fully implement dependency injection. That will probably be a goal of mine before I continue with U.I. design.